+++ caikit/runtime/http_server/http_server.py
@@ -46,6 +46,10 @@ from caikit.config import get_config
 from caikit.core.data_model import DataBase
 from caikit.core.data_model.dataobject import make_dataobject
 from caikit.core.toolkit.sync_to_async import async_wrap_iter
+from caikit.runtime.http_server.pydantic_wrapper import (
+    dataobject_to_pydantic,
+    pydantic_to_dataobject,
+)
 from caikit.runtime.server_base import RuntimeServerBase
 from caikit.runtime.service_factory import ServicePackage
 from caikit.runtime.service_generation.rpcs import (
@@ -56,10 +60,6 @@ from caikit.runtime.service_generation.rpcs import (
 from caikit.runtime.servicers.global_predict_servicer import GlobalPredictServicer
 from caikit.runtime.servicers.global_train_servicer import GlobalTrainServicer
 from caikit.runtime.types.caikit_runtime_exception import CaikitRuntimeException
-from caikit.runtime.wrappers.pydantic_wrapper import (
-    dataobject_to_pydantic,
-    pydantic_to_dataobject,
-)
 
 ## Globals #####################################################################
